@@ -5,12 +5,13 @@ import android.content.Context;
 import android.graphics.Color;
 import android.graphics.drawable.ColorDrawable;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
 import android.view.Window;
 import android.widget.ProgressBar;
 import android.widget.TextView;
 
 import com.yun.opernv2.R;
+
+import androidx.annotation.NonNull;
 
 
 /**
