@@ -1,15 +1,8 @@
 package com.yun.opernv2.ui.fragment;
 
 
-import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.widget.CardView;
-import android.support.v7.widget.DefaultItemAnimator;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -18,20 +11,19 @@ import android.widget.TextView;
 
 import com.bumptech.glide.Glide;
 import com.jakewharton.rxbinding2.view.RxView;
-import com.youth.banner.Banner;
-import com.youth.banner.BannerConfig;
-import com.youth.banner.Transformer;
-import com.youth.banner.loader.ImageLoader;
 import com.yun.opernv2.R;
-import com.yun.opernv2.model.OpernInfo;
+import com.yun.opernv2.model.ScoreBaseInfoDO;
 import com.yun.opernv2.net.HttpCore;
-import com.yun.opernv2.net.request.GetRandomOpernReq;
 import com.yun.opernv2.ui.activity.ShowImageActivity;
-import com.yun.opernv2.ui.activity.WebViewActivity;
 import com.yun.opernv2.utils.ToastUtil;
 
 import java.util.ArrayList;
 
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.DefaultItemAnimator;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import io.reactivex.android.schedulers.AndroidSchedulers;
@@ -46,10 +38,10 @@ public class HomeFragment extends Fragment {
     @BindView(R.id.empty_view)
     View emptyView;
 
-    private ArrayList<OpernInfo> opernInfoArrayList = new ArrayList<>();
+    private ArrayList<ScoreBaseInfoDO> scoreBaseInfoDOArrayList = new ArrayList<>();
     private Adapter adapter;
     private LinearLayoutManager linearLayoutManager;
-    private int index = 0;
+    private int page = 0;
     private boolean requesting = false;
 
     @Override
@@ -65,9 +57,7 @@ public class HomeFragment extends Fragment {
         linearLayoutManager = new LinearLayoutManager(getActivity());
         opernLv.setLayoutManager(linearLayoutManager);
         opernLv.setItemAnimator(new DefaultItemAnimator());
-        adapter = new Adapter(opernInfoArrayList);
-        View headerView = LayoutInflater.from(getContext()).inflate(R.layout.header_home_rv, null);
-        adapter.addHeaderView(headerView);
+        adapter = new Adapter(scoreBaseInfoDOArrayList);
         opernLv.setAdapter(adapter);
         opernLv.addOnScrollListener(new RecyclerView.OnScrollListener() {
             @Override
@@ -75,7 +65,7 @@ public class HomeFragment extends Fragment {
                 super.onScrolled(recyclerView, dx, dy);
                 if (dy > 0) {
                     int lastVisibleItem = linearLayoutManager.findLastVisibleItemPosition();
-                    int totalItemCount = opernInfoArrayList.size();
+                    int totalItemCount = scoreBaseInfoDOArrayList.size();
                     if (lastVisibleItem >= totalItemCount - 10) {
                         if (!requesting) {
                             net();
@@ -86,7 +76,7 @@ public class HomeFragment extends Fragment {
         });
         opernSrl.setColorSchemeColors(getResources().getColor(R.color.black));
         opernSrl.setOnRefreshListener(() -> {
-            index = 0;
+            page = 0;
             net();
         });
         net();
@@ -95,22 +85,21 @@ public class HomeFragment extends Fragment {
     private void net() {
         requesting = true;
         opernSrl.setRefreshing(true);
-        GetRandomOpernReq req = new GetRandomOpernReq();
-        req.setPageSize(40);
+
         HttpCore.getInstance().getApi()
-                .getRandomOpernInfo(req)
+                .homeRecommend(page)
                 .subscribeOn(new NewThreadScheduler())
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribe(arrayListBaseResponse -> {
-                    if (index == 0) {
-                        opernInfoArrayList.clear();
+                    if (page == 0) {
+                        scoreBaseInfoDOArrayList.clear();
                     }
-                    ArrayList<OpernInfo> data = arrayListBaseResponse.getData();
+                    ArrayList<ScoreBaseInfoDO> data = arrayListBaseResponse.getData();
                     if (data == null || data.size() == 0) {
                         ToastUtil.showShort("没有更多数据了");
                     } else {
-                        opernInfoArrayList.addAll(data);
-                        index++;
+                        scoreBaseInfoDOArrayList.addAll(data);
+                        page++;
                     }
                     adapter.notifyDataSetChanged();
                     opernSrl.setRefreshing(false);
@@ -123,113 +112,35 @@ public class HomeFragment extends Fragment {
                 });
     }
 
-
-    public class BannerImageLoader extends ImageLoader {
-
-        @Override
-        public void displayImage(Context context, Object path, ImageView imageView) {
-            Glide.with(context).load(path).into(imageView);
-        }
-
-        @Override
-        public ImageView createImageView(Context context) {
-            ImageView imageView = new ImageView(context);
-            imageView.setScaleType(ImageView.ScaleType.CENTER_CROP);
-            return imageView;
-        }
-    }
-
-
     public class Adapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
-        private static final int TYPE_HEADER = 0;
-        private static final int TYPE_NORMAL = 1;
-        private ArrayList<OpernInfo> opernInfoArrayList;
-        private View headerView;
-        private HeaderViewViewHolder headerViewViewHolder;
+        private ArrayList<ScoreBaseInfoDO> scoreBaseInfoDOArrayList;
 
-        public HeaderViewViewHolder getHeaderViewViewHolder() {
-            return headerViewViewHolder;
-        }
-
-        public void addHeaderView(View headerView) {
-            this.headerView = headerView;
-            notifyItemInserted(0);
-        }
-
-        public View getHeaderView() {
-            return headerView;
-        }
-
-        public int getRealPosition(RecyclerView.ViewHolder holder) {
-            int position = holder.getLayoutPosition();
-            return headerView == null ? position : position - 1;
-        }
-
-        public int getRealPosition(int position) {
-            return headerView == null ? position : position - 1;
-        }
-
-
-        public Adapter(ArrayList<OpernInfo> opernInfoArrayList) {
-            this.opernInfoArrayList = opernInfoArrayList;
-        }
-
-        @Override
-        public int getItemViewType(int position) {
-            if (headerView == null) return TYPE_NORMAL;
-            if (position == 0) return TYPE_HEADER;
-            return TYPE_NORMAL;
+        public Adapter(ArrayList<ScoreBaseInfoDO> scoreBaseInfoDOArrayList) {
+            this.scoreBaseInfoDOArrayList = scoreBaseInfoDOArrayList;
         }
 
         @Override
         public RecyclerView.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
-            if (viewType == TYPE_HEADER) {
-                return headerViewViewHolder = new HeaderViewViewHolder(headerView);
-            }
             return new ViewHolder(getActivity().getLayoutInflater().inflate(R.layout.item_opern_list, parent, false));
         }
 
         @Override
         public void onBindViewHolder(RecyclerView.ViewHolder viewHolder, int position) {
-            if (getItemViewType(position) == TYPE_HEADER) {
-                HeaderViewViewHolder holder = (HeaderViewViewHolder) viewHolder;
-                holder.banner.setBannerStyle(BannerConfig.CIRCLE_INDICATOR);
-                holder.banner.setImageLoader(new BannerImageLoader());
-                ArrayList<String> images = new ArrayList<>();
-                images.add(HttpCore.BaseUrl + "resources/image/ic_my_website.png");
-                holder.banner.setImages(images);
-                holder.banner.setBannerAnimation(Transformer.Default);
-                holder.banner.isAutoPlay(true);
-                holder.banner.setDelayTime(4000);
-                holder.banner.setIndicatorGravity(BannerConfig.CENTER);
-                holder.banner.start();
-                holder.banner.setOnBannerListener(position1 -> {
-                    Intent intent = new Intent(getContext(), WebViewActivity.class);
-                    switch (position1) {
-                        case 0:
-                            intent.putExtra("url", "http://60.205.182.130:8080/OpernServer/");
-                            break;
-                    }
-                    startActivity(intent);
-                });
-            } else {
-                ViewHolder holder = (ViewHolder) viewHolder;
-                OpernInfo opernInfo = opernInfoArrayList.get(getRealPosition(position));
-                Glide.with(getContext()).load(opernInfo.getOpernFirstPicUrl()).into(holder.opernImg);
-                holder.titleTv.setText(opernInfo.getOpernName());
-                holder.wordAuthorTv.setText("作词：" + opernInfo.getOpernWordAuthor());
-                holder.songAuthorTv.setText("作曲：" + opernInfo.getOpernSongAuthor());
-                holder.dataOriginTv.setText(opernInfo.getOriginName());
-
-            }
+            ViewHolder holder = (ViewHolder) viewHolder;
+            ScoreBaseInfoDO scoreBaseInfoDO = scoreBaseInfoDOArrayList.get(position);
+            Glide.with(getContext()).load(scoreBaseInfoDO.getScoreCoverPicture()).into(holder.opernImg);
+            holder.titleTv.setText(scoreBaseInfoDO.getScoreName());
+            holder.wordAuthorTv.setText("作词：" + scoreBaseInfoDO.getScoreWordWriter());
+            holder.songAuthorTv.setText("作曲：" + scoreBaseInfoDO.getScoreSongWriter());
+            holder.dataOriginTv.setText(scoreBaseInfoDO.getScoreOrigin());
+            holder.categoryTv.setText(scoreBaseInfoDO.getScoreFormat());
         }
 
         @Override
         public int getItemCount() {
-            emptyView.setVisibility(opernInfoArrayList.size() == 0 ? View.VISIBLE : View.GONE);
-            return headerView == null ? opernInfoArrayList.size() : opernInfoArrayList.size() + 1;
+            emptyView.setVisibility(scoreBaseInfoDOArrayList.size() == 0 ? View.VISIBLE : View.GONE);
+            return scoreBaseInfoDOArrayList == null ? 0 : scoreBaseInfoDOArrayList.size();
         }
-
 
         class ViewHolder extends RecyclerView.ViewHolder {
             @BindView(R.id.item_opern_list_img)
@@ -250,29 +161,12 @@ public class HomeFragment extends Fragment {
                 ButterKnife.bind(this, itemView);
                 RxView.clicks(itemView).subscribe(o -> {
                     Intent intent = new Intent(getActivity(), ShowImageActivity.class);
-                    intent.putExtra("opernInfo", opernInfoArrayList.get(getRealPosition(this)));
+                    intent.putExtra("scoreBaseInfoDO", scoreBaseInfoDOArrayList.get(getAdapterPosition()));
                     startActivity(intent);
                 });
             }
         }
 
-        class HeaderViewViewHolder extends RecyclerView.ViewHolder {
-            @BindView(R.id.banner)
-            Banner banner;
-            @BindView(R.id.lastUpdateCardView)
-            CardView lastUpdateCardView;
-            @BindView(R.id.lastUpdateTimeTv)
-            TextView lastUpdateTimeTv;
-            @BindView(R.id.musicChart_cardview)
-            CardView musicChartCardView;
-
-            public HeaderViewViewHolder(View itemView) {
-                super(itemView);
-                ButterKnife.bind(this, itemView);
-                RxView.clicks(lastUpdateCardView).subscribe(o -> ToastUtil.showShort("暂未开放"));
-                RxView.clicks(musicChartCardView).subscribe(o -> ToastUtil.showShort("暂未开放"));
-            }
-        }
     }
 
 }
