@@ -8,7 +8,6 @@ import android.content.res.Configuration;
 import android.content.res.Resources;
 import android.graphics.Bitmap;
 import android.provider.Settings;
-import android.support.annotation.NonNull;
 import android.util.DisplayMetrics;
 import android.view.Surface;
 import android.view.View;
@@ -16,6 +15,8 @@ import android.view.Window;
 import android.view.WindowManager;
 
 import com.yun.opernv2.Application;
+
+import androidx.annotation.NonNull;
 
 
 public final class ScreenUtils {
