@@ -1,9 +1,10 @@
 package com.yun.opernv2.views;
 
 import android.content.Context;
-import android.support.v4.view.ViewPager;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
+
+import androidx.viewpager.widget.ViewPager;
 
 /**
  * Created by Yun on 2017/9/18 0018.
